requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pypdf>=3.17.0
faiss-cpu>=1.7.4
chromadb>=0.4.18
redis>=5.0.1
python-multipart>=0.0.6
pydantic>=2.4.2
twilio>=8.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
reportlab>=4.0.4 